+++ tests/testthat/test-req-perform-parallel.R
@@ -3,9 +3,15 @@ test_that("request and paths must match", {
   expect_snapshot(req_perform_parallel(req, letters), error = TRUE)
 })
 
-test_that("correctly prepares request", {
-  reqs <- list(request_test("/post") %>% req_method("POST"))
-  expect_no_error(req_perform_parallel(reqs))
+test_that("can perform zero requests", {
+  expect_equal(req_perform_parallel(list()), list())
+})
+
+test_that("can perform a single request", {
+  reqs <- list(request_test("/get"))
+  resps <- req_perform_parallel(reqs)
+  expect_type(resps, "list")
+  expect_length(resps, 1)
 })
 
 test_that("requests happen in parallel", {
@@ -25,7 +31,7 @@ test_that("requests happen in parallel", {
 test_that("can perform >128 file uploads in parallel", {
   temp <- withr::local_tempfile(lines = letters)
   req <- request(example_url()) %>% req_body_file(temp)
-  reqs <- rep(list(req), 150)
+  reqs <- rep(list(req), 130)
 
   expect_no_error(req_perform_parallel(reqs, on_error = "continue"))
 })
@@ -111,7 +117,7 @@ test_that("errors can cancel outstanding requests", {
     request_test("/status/:status", status = 404),
     request_test("/delay/:secs", secs = 2),
   )
-  out <- req_perform_parallel(reqs, on_error = "return")
+  out <- req_perform_parallel(reqs, on_error = "return", max_active = 1)
   expect_s3_class(out[[1]], "httr2_http_404")
   expect_null(out[[2]])
 })
@@ -127,7 +133,6 @@ test_that("req_perform_parallel resspects http_error() error override", {
   expect_equal(resp_status(resps[[2]]), 500)
 })
 
-
 test_that("req_perform_parallel respects http_error() body message", {
   reqs <- list2(
     req_error(request_test("/status/:status", status = 404), body = ~"hello")
@@ -135,6 +140,187 @@ test_that("req_perform_parallel respects http_error() body message", {
   expect_snapshot(req_perform_parallel(reqs), error = TRUE)
 })
 
+test_that("respects max retries", {
+  req <- local_app_request(function(req, res) {
+    i <- res$app$locals$i %||% 1
+    res$
+      set_status(429)$
+      set_header("retry-after", 0)$
+      send_json(list(status = "waiting"), auto_unbox = TRUE)
+  })
+  req <- req_retry(req, max_tries = 3)
+  queue <- RequestQueue$new(list(req), progress = FALSE)
+
+  queue$process()
+  expect_s3_class(queue$resps[[1]], "httr2_http_429")
+  expect_equal(queue$tries[1], 3)
+})
+
+test_that("requests are throttled", {
+  withr::defer(throttle_reset())
+
+  mock_time <- 0
+  local_mocked_bindings(
+    unix_time = function() mock_time,
+    Sys.sleep = function(seconds) mock_time <<- mock_time + seconds
+  )
+
+  req <- request_test("/status/:status", status = 200)
+  req <- req %>% req_throttle(capacity = 1, fill_time_s = 1)
+  reqs <- rep(list(req), 5)
+
+  queue <- RequestQueue$new(reqs, progress = FALSE)
+  queue$process()
+  expect_equal(mock_time, 4)
+})
+
+
+# Tests of lower-level operation -----------------------------------------------
+
+test_that("can retry an OAuth failure", {
+  req <- local_app_request(function(req, res) {
+    i <- res$app$locals$i %||% 1
+    if (i == 1) {
+      res$app$locals$i <- 2
+      res$
+        set_status(401)$
+        set_header("WWW-Authenticate", 'Bearer realm="example", error="invalid_token"')$
+        send_json(list(status = "failed"), auto_unbox = TRUE)
+    } else {
+      res$send_json(list(status = "done"), auto_unbox = TRUE)
+    }
+  })
+  req <- req_policies(req, auth_oauth = TRUE)
+
+  reset <- 0
+  local_mocked_bindings(req_auth_clear_cache = function(...) reset <<- reset + 1)
+
+  queue <- RequestQueue$new(list(req), progress = FALSE)
+  queue$process()
+
+  expect_equal(reset, 1)
+  expect_equal(resp_body_json(queue$resps[[1]]), list(status = "done"))
+})
+
+test_that("but multiple failures causes an error", {
+  req <- local_app_request(function(req, res) {
+    res$
+      set_status(401)$
+      set_header("WWW-Authenticate", 'Bearer realm="example", error="invalid_token"')$
+      send_json(list(status = "failed"), auto_unbox = TRUE)
+  })
+  req <- req_policies(req, auth_oauth = TRUE)
+
+  queue <- RequestQueue$new(list(req), progress = FALSE)
+  queue$process()
+  expect_s3_class(queue$resps[[1]], "httr2_http_401")
+})
+
+test_that("can retry a transient error", {
+  req <- local_app_request(function(req, res) {
+    i <- res$app$locals$i %||% 1
+    if (i == 1) {
+      res$app$locals$i <- 2
+      res$
+        set_status(429)$
+        set_header("retry-after", 2)$
+        send_json(list(status = "waiting"), auto_unbox = TRUE)
+    } else {
+      res$send_json(list(status = "done"), auto_unbox = TRUE)
+    }
+  })
+  req <- req_retry(req, max_tries = 2)
+
+  mock_time <- 1
+  local_mocked_bindings(
+    unix_time = function() mock_time,
+    Sys.sleep = function(seconds) mock_time <<- mock_time + seconds
+  )
+
+  queue <- RequestQueue$new(list(req), progress = FALSE)
+
+  # Start processing
+  expect_null(queue$process1())
+  expect_equal(queue$queue_status, "working")
+  expect_equal(queue$n_active, 1)
+  expect_equal(queue$n_pending, 0)
+
+  # No pending, so switch to finishing
+  expect_null(queue$process1())
+  expect_equal(queue$queue_status, "finishing")
+
+  # Now we process the request and capture the retry
+  expect_null(queue$process1())
+  expect_equal(queue$queue_status, "working")
+  expect_equal(queue$rate_limit_deadline, mock_time + 2)
+  expect_equal(queue$n_pending, 1)
+  expect_s3_class(queue$resps[[1]], "httr2_http_429")
+  expect_equal(resp_body_json(queue$resps[[1]]$resp), list(status = "waiting"))
+
+  # Now we "wait" 2 seconds
+  expect_null(queue$process1())
+  expect_equal(queue$queue_status, "working")
+  expect_equal(mock_time, 3)
+
+  # Then resume finishing again
+  expect_null(queue$process1())
+  expect_equal(queue$queue_status, "finishing")
+
+  # And we're finally done
+  expect_null(queue$process1())
+  expect_equal(queue$queue_status, "done")
+  expect_false(queue$process1())
+
+  expect_equal(resp_body_json(queue$resps[[1]]), list(status = "done"))
+})
+
+test_that("throttling is limited by deadline", {
+  withr::defer(throttle_reset("test"))
+
+  mock_time <- 0
+  local_mocked_bindings(
+    unix_time = function() mock_time,
+    Sys.sleep = function(seconds) mock_time <<- mock_time + seconds
+  )
+
+  req <- request_test("/status/:status", status = 200)
+  req <- req_throttle(req, capacity = 1, fill_time_s = 1, realm = "test")
+  queue <- RequestQueue$new(list(req), progress = FALSE)
+
+  # Check time only advances by one second, and token is returned to bucket
+  local_mocked_bindings(throttle_deadline = function(...) mock_time + 2)
+  queue$submit_next(1)
+  expect_equal(mock_time, 1)
+  expect_equal(the$throttle[["test"]]$tokens, 1)
+
+  local_mocked_bindings(throttle_deadline = function(...) mock_time)
+  queue$rate_limit_deadline <- mock_time + 2
+  expect_equal(mock_time, 1)
+  expect_equal(the$throttle[["test"]]$tokens, 1)
+})
+
+# Pool helpers ----------------------------------------------------------------
+
+test_that("wait for deadline waits after pool complete", {
+  pool <- curl::new_pool()
+  deadline <- unix_time() + 1
+
+  slept <- 0
+  local_mocked_bindings(
+    unix_time = function() 0,
+    Sys.sleep = function(seconds) mock_time <<- slept <<- seconds
+  )
+
+  expect_true(pool_wait_for_deadline(pool, deadline = 1))
+  expect_equal(slept, 1)
+})
+
+# Deprecations ----------------------------------------------------------------
+
 test_that("multi_req_perform is deprecated", {
   expect_snapshot(multi_req_perform(list()))
+})
+
+test_that("pool argument is deprecated", {
+  expect_snapshot(. <- req_perform_parallel(list(), pool = curl::new_pool()))
 })